more fine-grained ts targets
there are now per-language ts-<package>-<lang> and ts-<lang> targets to
update only the files relevant for a given language.
there is also a ts-all target which does what the old ts target did.
the ts-<package> targets are now named ts-<package>-all and should not
be used manually - they exist for the convenience of the ts-all target.
the ts target is now only a help blurb.