Ensure that the view loses its selection when typing in the lineedit
When QT_NO_FSCOMPLETER is defined then when selecting a file in the view
in a filedialog and then changing the filename in the lineedit to a file
that does not exist then the view should deselect any selection it has.

The name of the slot in this case is unfortunate as it does not actually
do any auto-completion.  This ensures that the same behaviour works in
both cases.

Reviewed-by: Alexis Menard