fix for looping crash log on data abort
When a thread panics, calling resume allows the thread to proceed to the
exit handler and shut down.
When a thread takes an exception, calling resume continues from the same
point so the exception happens again.
To avoid this, we add crashed thread ids to a list. If we see the same
thread crash again, we terminate the process without trying to fetch the
call stack a second time.

Reviewed-by: Jason Barron