QWorkspace: fix hardcoded min size overwriting the real min size
The setMinimumSize call was having precedence over the minimumSizeHint,
so when having a sizegrip enabled it was possible to resize a MDI window to
a much smaller size than its min size (actually it would flicker between
the two sizes, on each resize). Fixed by moving that min size (for the
titlebar contents) to the calculations in minimumSizeHint().

Reviewed-by: mariusso