Windows mobile: the [X] button in the taskbar minimizes the widget
The [X] or cancel button in the task bar shall just "minimize" the
widget on Windows mobile. A press on this button results in a
WM_COMMAND, IDCANCEL message. Before this patch we just sent a
QCloseEvent to the widget, which had basically no effect.
Now, we're calling showMinimzed(), which is the desired behaviour.

Task-number: QTBUG-8408
Reviewed-by: Martin Petersson