Used aligned load and store when possible for the blending of ARGB32
Unaligned load and store can be costly. This patch mitigate the problem
by aligning the destination before using SSE2. The destination is
aligned because it is used by load and store, while the source is only
use by load.

On Atom, the blending test is 7% faster for ARGB32.

Re-pushing that patch, thanks to awesome policies...

Reviewed-by: Andreas Kling