Fixed a QSplashScreen hanging bug in S60 3.1 devices.
QSymbianBitmapDataAccess is used to provide access to the bitmap heap
in a manner that locks correctly on all platform versions. The
heapWasLocked variable was meant to protect against the case where
the heap is locked recursively. However, it failed to take into
account the case where the same QSymbianBitmapDataAccess object was
used to lock recursively. In this case the variable would be changed
to true on the second lock, which means that the lock would never be
released again.

This was fixed by making the access reference counted instead. Since
the bitmap heap lock is global, the refcount was made global as well.

Task:     QTBUG-11129
RevBy:    Jason Barron
AutoTest: Works again. It was hanging before this fix.
(cherry picked from commit d4cc1dcea5b4116767cfee0ec45bfba72dc011ff)