moc: Slot with complex template default value does not compile
The way we detect the end of a default argument does not take in account
template parametter.
It is unfortunatelly not trivial to do it properly without semantic information
So we will use heuristics and if the number of < matches the number of >
we consider it is a template.  Or if we have a '=' we consider it is not a
template.

Task-number: QTBUG-12260
Reviewed-by: Roberto Raggi