Improve QML text rendering when LCD smoothing is enabled for OS X.
Text in QML is always painted into a pixmap, and with sub-pixel
antialiasing enabled this produced incorrect results. This patch is a
temporary fix until QTBUG-12252 and/or QTBUG-7747 can be properly
addressed in a patch release. In the worst case, if using QML and
QPainters are being constructed in a non-GUI thread, this patch could
potentially turn off sub-pixel antialiasing for them.

Task-number: QTBUG-11002
Reviewed-by: Aaron Kennedy
Reviewed-by: Eskil Abrahamsen Blomfeldt