QGraphicsItem: Animation leaves drawing artifacts when clipping is used.
This only happens when the ItemHasNoContents and ItemClipsChildrenToShape
flags are set. Problem is that items with no content are threated as 'dummy'
items, which means they are never drawn or 'processed' otherwise, so the
cached bounding rect is not reliable/usable. This means that in case of
changing the geometry of such items, its children always have to take
care of invalidating the occupied areas and the update can not be
clipped to the item's bounding rect.

Regression after commit: c1c7dbf2

Auto test included.

Task-number: QTBUG-11504
Reviewed-by: yoann
(cherry picked from commit 4df9c96e2c213c39924e22e02621b0c61e83f8fe)