Do not crash due to a infinite recursion when using voiceover on MacOS
The reason for the infinite recursion was that
QAccessibleTitleBar::object() returned the titlebar's dockwidget.
This could lead to a problem when the AT client tried to
traverse the accessibility hierarchy:
As a response to QAXChildrenAttribute (retrieve children) on a dock
widget node, it would register it's children in the
hierarchy manager. In this case, the object registered for the titlebar
interface was the QDockWidget.
In order to do further traversal, the bridge could call
queryAccessibleInterface on the list of retrieved children to get the
QAccessibleInterface for those objects, however, that would return the
QAccessibleDockWidget interface that we just had traversed,....

Task-number: QTBUG-6843
Reviewed-by:  Carlos Manuel Duclos Vergara
(cherry picked from commit 64b9e63f4f9162c1af299b1355e84b0e616ad768)