Remove the masking when computing qAlpha()
After a bit shift of 24, only the alpha value should remain, so it is
not necessary to mask the result.
The documentation state QRgb works on a ARGB quadruplet, so the upper
bits can be assumed to be zero in the cases when QRgb is 64 bits.

This saves some time because qAlpha() is used for each pixel in the
generic blend functions.

Reviewed-by: Andreas Kling
Reviewed-by: Kim