Removed createBlittable factory function from PlatformIntegration
and moved it onto QBlittablePixmap making the class abstract. The
creator function was seen as to spesialised to be part of the
PlatformIntegration class, and belonged to the QBlittablePixmap class
anyway since that was the PixmapData type you would need to return to
use the blitter api. Also removed the cross platform examples as they
are more confusing than anything else. For usage of QBlittable look in
the directfb platform integration plugin.