Fix crash caused by not clearing the QPixmapCache on application exit.
The qt_cleanup() function will call CCoeEnv::DestroyEnvironment() on
application shutdown. This destruction will in turn tear down the
various server sessions associated with the application. One of these
sessions is the FBSERV session and if the QPixmapCache attempts to
delete a CFbsBitmap after the session has been destroyed, the app will
crash.

The solution is call QPixmapCache::cleanup() before we destroy the
environment. This is inline with what the other platforms do.

Reviewed-by: Jason Barron