Add MulticastTtlOption and MulticastLoopbackOption to QAbstractSocket
These options only make sense and work in QUdpSocket when using
multicasting. Previously, the loopback was set when joining a multicast
group, but the loopback flag is not a per-group flag, it's per socket,
hence this change.

The ability to set the multicast TTL is needed as well, so this is added
as a socket option as well.

Autotest updated to test both.