Report errors when evaluating JavaScript expressions from the GestureArea QML plugin.
Before the patch if the evaluation fails (for any reason - because you access
invalid or non-existing property, or because of a spelling mistake, etc), the
expression will be silently ignored, making if very hard to figure out what is
going wrong.

The patch makes sure the error description will be printed on the output.

Reviewed-by: Warwick Allison