Ensure backing store is deleted before top-level window
If this is not done, later deletion of the backing store may cause
a crash.  If the backing store is an EGL surface, its destruction
includes a call to eglDestroySurface(), which triggers an exception
if the window handle passed as a parameter is no longer valid.

Task-number: QTBUG-10643
Task-number: QTBUG-11376
Reviewed-by: Jason Barron