Clean the CompositionFunction tables of drawhelper
Some instructions sets were defining partial table of composition
functions. To work around that, qInitDrawhelperAsm() was resetting
the composition function of QPainter::CompositionMode_Destination and
anything above QPainter::CompositionMode_Xor.

This was a problem because it makes it impossible to implement fast
path for those composition mode.

This patch export prototypes for the generic functions of each
composition mode. The specialized implementations now define a complete
table.

Reviewed-by: Andreas Kling