Fixed the following sub-tasks for QTBUG-12192
* Qt Simulator link missing

* Integration and add-ins (on the front page, link leads to an alternative start page)

* Qt Creator (on the front page link leads to Manual version 1.3)

* Getting Started (on the front page) link broken

* More... (in Global Declarations) link broken

Additionally:

There was an issue with the documentation used within Assistant/Creator that appears to have no bug report filed. An alternate "offline" form of the documentation (inspired by the newly redesigned online version) was being installed, but had significant issues within Assistant and Creator. It appears that within qdoc3 some changes had been made towards using this newer documentation format, but those changes were not working properly.

I fixed qdoc3 so that the original offline documentation design is the default, but there is room to use the configuration files to call the any of the three documentation designs (new online, new offline, and old offline).