Workaround for QTBUG-8013: Do not return an ascent of 0
Symbian's CFont::FontMaxAscent() returns in some cases an
incorrect value of 0. That usually happens (for some font sizes)
if a stroke based font is the main system font.

We were able to reproduce it on some S60 3.2 devices with a
chinese language pack installed.

This patch will test if CFont::FontMaxAscent() returns 0. And if
so, it alculates an ascent taht makes more sense.

Task-number: QTBUG-8013
Reviewed-by: Liang Qi