Don't run the QGL test on systems that does not have GL support.
On a Mac that has no screens connected, it is possible to create
offscreen GL contexts, but not normal QGLWidgets. There are some
assumptions both in the tests and in the QGL module that this if you
can create one of them, you can create both.
The test now check if it can create a valid QGLWidget, if it can't
it will skip running all the QGL tests.

Task-number: QTBUG-12138
Reviewed-by: Prasanth