Fix signal emission of QDesktopWidget on Symbian.
QDesktopWidget on Symbian was not properly emitting the resized() and
the workAreaResized() signals. Qt was properly translating the system
event and sending the QResizeEvent to the widget, but the signal
emission is determined solely based on changes in the QVectors of
QRects outlining the screen dimensions. On Symbian, these vectors were
never being set and as a result the signals were never omitted.

The fix is to set the rects correctly in the init() function which is
called by the resizeEvent() handler. Also modify the accessor functions
to return the rectangles stored in the vector instead of querying for
them each time.

Task-number: QTBUG-11433
Reviewed-by: axis
Autotest: PASSES