Ensure that window rectangle is updated when CBA visibility changes
In S60, there are two pieces of screen furniture, changes in whose
visibility or size can affect the size of the Qt application window.
These are the status pane, at the top of the screen, and the CBA at
the bottom.

QSymbianControl listens for changes in status pane visibility and
size by implementing MEikStatusPaneObserver.  Notifications
received via this interface trigger a call to
QSymbianControl::handleClientAreaChange() which resizes the control.

There is no corresponding interface through which to receive
notifications of changes in CBA visibility.

This patch introduces a utility function for setting the visibility
of both the status pane and CBA, which ensures that the control's
rectangle is updated when either one changes.

Task-number: QTBUG-5320
Reviewed-by: Jason Barron
(cherry picked from commit 92d6d1c974206bac7520101ad26ca93eb0b4d977)