Wrong geometery for QMainWindow with Unified toolbar on Cocoa.
If a mainwindow is shown maximized (initially) with unified toolbar,
the top portion of the widget gets hidden underneath the native toolbar.
This is because while showing the window, the toolbar is resized. Which
causes the client area of the NSWindow to resize, we need to update the
frame of the NSView for QMainWindow based on this new size.

Task-number: QTBUG-10500
Reviewed-by: Denis