Fixed a problem with dso dependencies.
Dependencies to dso files were not generated correctly, both because
the PRE_TARGETDEPS variable seems to be local to the function it was
in, and also because the regular expression was wrong (it has to
match the whole string).

Also switched to qt_ style variable names to avoid clashes.

RevBy:    Trust me