Added support to Phonon MMF backend for playback of Qt resource files
The backend accesses the resource file path via MediaSource::url().
A small patch to Phonon was required to enable this, because by
default, Phonon passes a QIODevice, rather than the resource file
path, to the backend.

The backend uses this path to create a QResource object, through which
the memory buffer into which the resource file has been read can be
accessed.  This buffer is wrapped in a Symbian 8-bit descriptor and
passed to the OpenDesL() function of the appropriate MMF client
utility API.

Task-number: QTBUG-6562