Fix proxy widgets with the OpenVG paint engine.
Proxy widgets use the shared painter functionality and this implies
that the paint engine's begin function does not get called for each new
widget that gets painted. This causes a problem because the system clip
gets modified by QPainter, but the paint engine does not realize that
this happened and fails to use the new clip. The result is that you can
end up painting outside the intended clip area.

The fix is to reimplement the virtual systemStateChanged() function in
QVGPaintEnginePrivate and make it call updateScissor() to re-evaluate
the clipping flags and update the current clip accordingly.

A similar fix was done to the OpenGL paint engine way back in 307c2954.

Task-number: QTBUG-12486
Reviewed-by: Rhys Weatherley