Add benchmarks for QtScript's QObject binding
The benchmarks cover the following:
- Reading and writing meta-object-defined properties.
- Reading and writing dynamic properties.
- Reading child objects.
- Reading script-defined (wrapper object) properties.
- Reading prototype-inherited properties.
- Calling slots (overloaded and non-overloaded, with
  or without implicit argument conversion).
- Signal handling.

Both "simple" and "complex" objects are tested. A "simple" object
is an object that inherits directly from QObject, i.e. it doesn't
drag with it a lot of stuff. A "complex" object is an object that
has many properties and a relatively long inheritance chain
(currently QPushButton is used).

Reviewed-by: Jedrzej Nowacki