Switched from using hardlinks to copies in Symbian building.
The problem with hardlinks is that things start to behave funny when
you build several times, because on the next compile, the target from
armlink is moved to the .sym file location, and in this case it will
be hardlinked too. In the end you have all three files (bare, .exe
and .sym) being hardlinked together, when the .sym should be an ELF
file and the other two should be an E32 image.

RevBy:    Trust me