Cocoa: Active QDockWidget does not stay on top of inactive QDockWidget
If a window has several child windows, it was not be possible to
click on a child window to make it stack in front of the other
children. The reason is that cocoa held a fixed stacking order of
the child windows (sub-windows). This patch will release, and then
reestablish, the stacking order of the children upon acitvation.

Task-number: QTBUG-11780
Reviewed-by: prasanth