Use the fast conversion from RGB888 to RGB32 for Jpeg images
The commit 90642dd2b6b14c39cc6178f1161331895809b342 introduce a fast
way to convert from RGB888 to RGB32 with SSSE3.

This patch uses that function to do the color conversion when decoding
jpeg images.
The #defines for the SIMD extension have been moved to the common code
in order to be able to compile the jpeg handler directly in QtGui or
as a plugin.

Reviewed-by: Andreas Kling