Make SpringAnimation usable inside a transition
SpringFollow was turned into SpringAnimation so it could be
used inside a Behavior. This patch completes the work and makes
it usable inside a transition.

This is the documented behavior of SpringAnimation, the patch
is essential to make SpringAnimation work consistently.

Reviewed-by: Michael Brasser
Task-number: QTBUG-12141
(cherry picked from commit 33203efb786de9aebc6e05d1b4878d89e86ac456)