Use hasHeightForWidth() and hasWidthForHeight() instead of sizePolicy()
hasHeightForWidth() already does the same check (if the widget has a
layout, use the layouts size policy instead of the widgets').

Also generalized the implementation so that it now supports
width-for-height.
Also fixed a bug where we expected the other dimension to decrease when
the first dimension increased. This is not always the case.