Improving text coordinate rounding in the OpenVG paint engine
Commit b0bbabe728fedb8531fc2837403856bd5ed44e1b fixed text blurriness
in the OpenVG paint engine by forcing the coordinates of the text
items to integer coordinates (for unrotated, unscaled text).

That was not yet enough. In addition to the coordinates, also the
d->pathTransform in QVGPaintEngine can have a non-integer translation.

This patch makes sure that the text item coordinate combined with the
translation result in final integer coordinates.

Since it is not possible to set an absolute translation of a
QTransform (only relative is possible), first dx() is added to p.x(),
then after rounding, it is again substracted. Sam for y, but with
opposite prefix, since the y-axis in Qt and in transformation
matrices are in opposite directions.

The ceil stunt (which I cerated by trial and error) was replaced
by floor(x + aliasedCoordinateDelta), which *exactly* what other
paint engines do.

Task-number: QTBUG-12330
Reviewed-by: Jason Barron