Revert "Doc: Adding GS QML example files"
This reverts commit f13ebf52a0d091c6c72c97d2a13311e1e7c4a5cf.

The given example does not meet the autotest requirements for QML
examples. These involve that it starts without error messages, and that
it follows our naming convention of the start file starting with a
lowercase letter.

Please run the tests/auto/declarative/examples before pushing new
declarative examples. If new examples should be exempted from these
checks, add them to the list of exceptions in that autotest.