Fix IPv6 support on Windows
We need to make sure that we use the right setsockopt level and option
for the multicast TTL and loopback options. Like on UNIX, we need to use
IPPROTO_IPV6 level when joining and leaving a group. Also, we cannot
bind to a multicast address on Windows (but it is possible on UNIX).
work around this by binding to the Any/AnyIPv6 addresses if a multicast
address is passed to bind.