Rename Symbian generated mmp/mk files to include target in filename
Mmp and mk files previously contained UID3 as unique identifier in
the filename. However, this was not particularly useful for building
subprojects in massive project trees such as Qt, as it was difficult
to remember UID for each subproject. Now with intuitive naming of mmps,
it is possible to build a subproject like this:

sbs -c armv6_urel -p qtcore_dll.mmp

Task-number: QTBUG-12715
Reviewed-by: Jason Barron