Fixes for combining multiple Qt::ImhXXXOnly with S60 FEP
When one restriction is a superset of another (e.g. ImhDigitsOnly |
    ImhDialableCharactersOnly), the less restrictive one is used
When two restrictions are incompatible (e.g. ImhDialableCharactersOnly |
    ImhFormattedNumbersOnly), fall back to allowing all symbols

Note for some combinations additional characters not in the union can
be entered, this is a limitation of the API to AVKON FEP.
Before this change, some characters in the union could not be entered
at all, which is worse.

Tested on I8510 (3.2), 5800XM (5.0), N8(symbian^3)

Task Number: QTBUG-12726

Reviewed-by: Alessandro Portale