Fix scrollbar randomly popping up in QPlainTextEdit
When a line break was detected, we would retain the value of the
right bearing for the character after the break instead of resetting it
to the right bearing of the previous value. This could in some cases
cause the bounding rect of the text to be wrong, and could cause
unnecessary horizontal scrollbars to pop up. It was especially visible
when using WrapAnywhere. Visible e.g. in the compile output in Creator.

Done-by: mae
Reviewed-by: Eskil
Reviewed-by: Lars