Fix how subpixel positions are intepreted in an aliased grid.
With this change we break with the traditional X11 and GDI way of rasterzing
aliased polygons, but we become consistent across integer and sub-pixel
positions.

The new model uses the same rules as the antialised rasterizer which means
that a primitive edge on 0.4 becomes 0.0, 0.5 becomes 1 and 0.9 becomes one, which
is also in line with how one normally thinks of rounding rules.