qmake vcproj generator cleanup
MSVCPROJ_CXXFLAGS removed. This was unused.

MSVCPROJ_LFLAGS removed. This was used to convert the values of
QMAKE_LFLAGS, QMAKE_LIBDIR and DEF_FILE to linker command line options.
Then they were converted back to set values in the object model.
This journey to the world of command line options isn't necessary...

Reviewed-by: ossi