Fix memory leak in QtScript variable object
The d-pointer is of type JSVariableObjectData*, but
JSVariableObjectData doesn't have a virtual destructor.
Hence we must cast the d-pointer to our subclass when deleting.

In particular, this will ensure that the destructor of the
symbolTable member is called, which will deallocate the table
storage. (For QScriptActivationObject this did not cause a leak
in practice, because its symbolTable is always empty, and JSC's
hash table uses lazy allocation.)

Task-number: QTBUG-12479
Reviewed-by: Olivier Goffart