Replaced backing store reference count with list of visible widgets
Previously, the following sequence:
  1. Widget is hidden
  2. Widget partially revealed
  3. Widget fully revealed
resulted in the reference count of the backing store owned by the
widget's window() being incremented twice.

This patch replaces the simple reference count with a QSet which
stores pointers to the native widgets which are descendents of
the backing store owner, and which are currently visible.  The
sequence above therefore results in just a single insertion at
step (2), with step (3) having no effect on the backing store.

The QRefCountedWidgetBackingStore class has been renamed
QWidgetBackingStoreTracker to better reflect its purpose.

Task-number: QTBUG-12800
Task-number: QTBUG-12817
Reviewed-by: Jason Barron