Layout items had the wrong size if the layout was resized to maximum.
The problem was that if targetSize was the layouts maximum size,
then sumAvailable could be reduced to exactly 0 in the "while
(keepGoing)" loop. That would mean that we would not enter the block
that actually copied the new sizes from the temporary "newSizes" array
to the output "sizes" array due to the "if (sumAvailable > 0)"
condition.

Task-number: none, discovered while fixing QT-3570
Reviewed-by: Frederik Gladhorn