Disable minRightBearing optimization in QTextLayout on Mac
Neither of the Mac font engines have implemented minimum right bearing,
which will cause them to return 0 for the minimum value. This will
cause the right bearing to never be calculated prior to breaking
and thus never be a part of the breaking width. Since actually
implementing the functions is too time consuming for right now, we
will disable the optimization for the time being.

Reviewed-by: Olivier
(cherry picked from commit bb696453f6f87cad58982af3455335478fedb8cc)