Crash in QX11PaintEngine::drawPixmap
We receive a pixmap as a const reference and then we convert it to an
X11 pixmap. This conversion could fail for many reasons, however we
were not looking at the result of this conversion.
This patch was contributed by Christoph Feck from KDE.

Task-number: QTBUG-12826
Reviewed-by: Samuel
(cherry picked from commit 8e7fb343372acd0bc5fe04db7a5c5542541d4538)