Fix memory leak.
Mac only bug - run examples/opengl/overpainting under leak detect
instrument on Mac OSX to see it.

Memory leak in any application that called getProcAddress() - CFStrings
have to be cleaned up if the name of the function you got them from
contains the word "Create" - and we weren't doing that.

Reviewed-by: Gunnar Sletta