Introduce a second compatibility build key to Qt.
Some compilers are compatible with one another. In particular, the
Intel CC on Unix systems is compatible with g++ on the same system.

We should be saving the ABI "name" in the build key, not the compiler
name, but it's too late for that. Choose "g++-{VERSION}" as the
standard ABI name and make ICC support that.

Reviewed-by: Bradley T. Hughes