Revert the addition of the test of QtTest's random feature.
It was added in commit 56396b4b2b66da3133aff1b3864c550ce892d233.

Even though we have a -seed argument to provide reproduceability, it's
only reproduceable if the PRNG algorithm is also kept. So we can't
autotest this when we don't control the algorithm (for example, it's
different from glibc to MSVC's runtime).