Clean up rasterfallback mechanism in DirectFB
Being able to debug which paint operations are slow with Qt/DirectFB is
currently only possible in debug builds (which defeats the purpose
somewhat) or by frequent recompilation (with different defines).

This patch cleans up the mechanism somewhat and makes it available in
release builds as well.

The overhead is a single int comparison so it really isn't a performance
issue to have this code in place at all times.

Reviewed-by: Anders Bakken <agbakken@gmail.com>