diagramscene example: fix leak and crashes.
In MainWindow::deleteItem, if there is arrow selected, we need to clean
up the items from this arrow. We need to delete them first so there is no
arrow in the list anymore when we calls removeArrows (which before,
deleted arrow that were possibly in the list, resulting in crashes)

Also avoid leak by giving parents to objects that needs it, and destroying items.

Reviewed-by: Geir Vattekar
Task-number: QTBUG-12753