Fix check to remove unnecessary deployments in Symbian
Previously check to remove files from deployment in case the deployment
target location was the same as the source location was done using
QFileInfo equality, which resolved to true when file didn't yet exist
at the time qmake was run. Changed the check to compare absolute paths
instead.

Task-number: QTBUG-12879
Reviewed-by: Shane Kearns