Multiple "Edit/Special Characters" menu entries on Mac OS X (Cocoa)
Every time [NSApp setMainMenu:] is called, Cocoa will add the 'Special
Characters' item to the 'Edit' menu. Before adding a new entry it will
make sure that menu items list doesn't contain an item with the selector
'orderFrontCharacterPalette' & a 'nil' target. We need to return the
index for the first entry (we have QCocoaMenuLoader as target).

Task-number: QTBUG-12842
Reviewed-by: Denis