Cocoa: Stacking order of modal dialogs is wrong
Modal dialogs were sometimes hidden behind other normal windows
on screen. The reason was that, upon going modal for a window, we
actually resat the stacking level on the window in a faulty
attempt to respect any stays-on-top window flags. This patch
makes sure we avoid doing that, and at the same time, ensures
we don't reintroduce the original bug as well.

Task-number: QTBUG-12841
Reviewed-by: cduclos