Fix OpenGLVersionFlags on OpenGL 4.0 systems
If the OpenGL version was 4.0 or higher, the flags for 1.1, 1.5, 2.0,
and so on were not being set.  This lead to the flags indicating that
older OpenGL versions were not supported when they clearly were.

Back port parts of 28659c21d12a267b10e5b441bf4c776e04d69bdc from Qt 4.7
to fix this in Qt 4.6.

Task-number: QTBUG-12554
Reviewed-by: Sarah Smith