Cocoa: parent windows shows on screen when they should be hidden
The reason is the cocoa method that adds a window as a child of another,
it ends up showing both the parent and the child window regardless.
And this could in some cases also cause a crash.
So we therefore need to be careful when calling that function, and be
sure that the parent is actually visible. In addition, addChildWindow
reset the stacking level of the child window, and made e.g. normal
child windows pop in front of tool child windows. This could easily
be seen in e.g. Designer.

Task-number: QTBUG-12866
Reviewed-by: prasanth