Prevented Xorg crash in qtdemo when running corkboards example.
The crash happens in the Nvidia driver in glXReleaseTexImageEXT when
scrolling the corkboard using the mouse. To work around it we detect the
Nvidia driver versions where this is known to be a problem and skip
using the texture from pixmap extension in those cases.

Task-number: QTBUG-12914
Reviewed-by: Trond
(cherry picked from commit a76b8bf67696ae69888cc6237417e7c8f07f8da6)