QDeclarativeImageProvider: Do not keep the global declarative mutex locked when processing.
The point is to be able to process images in a thread. If the mutex
is locked, this is useless.

Use case is a slow QDeclarativeImageProvider that generates thumbmails
from large files. Even with the asynchronous attribute set to true, the
gui thread would be blocked by the mutex.

By using QSharedPointer, I also fix the leak of the providers (which
were not deleted)

Reviewed-by: Martin Jones